Refactor storage to have top-level mod.rs file
