+++ src/app/storage/mod.rs
@@ -1,0 +1,4 @@
+mod data_format;
+mod storage_service;
+
+pub use storage_service::StorageService;
